refactor: 타입별 extension 파일 분리
