+++ Calculator/Calculator/Model/Extension/Double.swift
@@ -1,0 +1,6 @@
+//
+//  Double.swift
+//  Calculator
+//
+
+extension Double: CalculateItem {}
